OCR/requirements.txt
absl-py==1.3.0
anyio==3.6.2
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
DateTime==4.7
decorator==5.1.1
easyocr==1.6.2
editdistance==0.6.1
efficientnet==1.0.0
essential-generators==1.0
fastapi==0.87.0
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
imageio==2.22.4
imgaug==0.4.0
keras==2.11.0
Keras-Applications==1.0.8
keras-ocr==0.8.9
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
networkx==2.8.8
ninja==1.11.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pathlib==1.0.1
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pydantic==1.10.2
pyparsing==3.0.9
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.1
tifffile==2022.10.10
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
validators==0.20.0
Werkzeug==2.2.2
wrapt==1.14.1
zope.interface==5.5.2
uvicorn[standard]
matplotlib
pillow
tesseract
fastapi_offline